istextfile: move to dvc.objects
