+++ dvc/objects/istextfile.py
